copy/paste split
